Implemented base of settings
Learning more rust in the process... thank you once_cell for making it possible to have a hash map in the static struct of the controller, also the compiler for recommending that.

Note: adding settings returns the default value, or the GUI value when the user changes it, so the developer is the one in charge of storing that value. I am doing it in a hash map.